// src/app/lib/enums/product.enum.ts
export enum EngineCapacity {
  SMALL = "125",
  NORMAL = "250",
  LARGE = "500",
  EXTRA_LARGE = "1000",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  SportBikes = "SportBikes",
  TouringBikes = "TouringBikes",
  CafeRacerBikes = "CafeRacerBikes",
  ElectricMotorcycles = "ElectricMotorcycles",
  Scooters = "Scooters",
}